@@ -1,5 +1,3 @@
-from typing import Optional
-
 from asyncpg import Record
 
 from udata_hydra import context
@@ -14,7 +12,7 @@ class Check:
     """Represents a check in the "checks" DB table"""
 
     @classmethod
-    async def get_by_id(cls, check_id: int, with_deleted: bool = False) -> Optional[Record]:
+    async def get_by_id(cls, check_id: int, with_deleted: bool = False) -> Record | None:
         pool = await context.pool()
         async with pool.acquire() as connection:
             q = """
@@ -29,7 +27,7 @@ class Check:
     @classmethod
     async def get_by_resource_id(
         cls, resource_id: str, with_deleted: bool = False
-    ) -> Optional[Record]:
+    ) -> Record | None:
         pool = await context.pool()
         async with pool.acquire() as connection:
             q = """
@@ -43,8 +41,8 @@ class Check:
 
     @classmethod
     async def get_latest(
-        cls, url: Optional[str] = None, resource_id: Optional[str] = None
-    ) -> Optional[Record]:
+        cls, url: str | None = None, resource_id: str | None = None
+    ) -> Record | None:
         column: str = "url" if url else "resource_id"
         pool = await context.pool()
         async with pool.acquire() as connection:
@@ -58,9 +56,7 @@ class Check:
             return await connection.fetchrow(q, url or resource_id)
 
     @classmethod
-    async def get_all(
-        cls, url: Optional[str] = None, resource_id: Optional[str] = None
-    ) -> Optional[list]:
+    async def get_all(cls, url: str | None = None, resource_id: str | None = None) -> list | None:
         column: str = "url" if url else "resource_id"
         pool = await context.pool()
         async with pool.acquire() as connection:
